// src/lib2/models/ApiResponseWrapperBudgetDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { BudgetDto } from './BudgetDto';
export type ApiResponseWrapperBudgetDto = {
    success?: boolean;
    message?: string;
    data?: BudgetDto;
    timestamp?: string;
    traceId?: string;
    path?: string;
    code?: number;
};