// src/lib2/models/ApiResponseWrapperBudgetVsRealiseDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { BudgetVsRealiseDto } from './BudgetVsRealiseDto';
export type ApiResponseWrapperBudgetVsRealiseDto = {
    success?: boolean;
    message?: string;
    data?: BudgetVsRealiseDto;
    timestamp?: string;
    traceId?: string;
    path?: string;
    code?: number;
};