// src/lib2/models/ApiResponseWrapperListBudgetDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { BudgetDto } from './BudgetDto';
export type ApiResponseWrapperListBudgetDto = {
    success?: boolean;
    message?: string;
    data?: Array<BudgetDto>;
    timestamp?: string;
    traceId?: string;
    path?: string;
    code?: number;
};