// src/lib2/models/ApiResponseWrapperVoid.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ApiResponseWrapperVoid = {
    success?: boolean;
    message?: string;
    data?: Record<string, any>;
    timestamp?: string;
    traceId?: string;
    path?: string;
    code?: number;
};